requirements.txt
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
colorama==0.4.6
cryptography==43.0.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
h11==0.14.0
hvac==2.3.0
idna==3.4
lxml==5.3.0
msal==1.30.0
numpy==1.26.0
Office365-REST-Python-Client==2.5.12
openpyxl==3.1.2
outcome==1.2.0
packaging==23.2
pandas==2.0.3
pycparser==2.21
PyJWT==2.9.0
PySocks==1.7.1
pyspnego==0.11.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.2
requests==2.31.0
requests-toolbelt==1.0.0
requests_ntlm==1.3.0
selenium==4.5.0
SharePlum==0.5.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sspilib==0.1.0
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2023.3
urllib3==1.26.17
webdriver-manager==4.0.1
wsproto==1.2.0
